Working search with instantsearch.js for flore & vegetation
