+++ modules/search-hit/search-hit-mixin.pug
@@ -12,17 +12,24 @@ mixin search-hit(data)
 				+search-hit-vegetation(data)
 
 mixin search-hit-flore(data)
-	- var title = data._highlightResult.bdtfx.scientific_name.value
-	- var text = data._highlightResult.bdtfx.common_name.value
+	//- bdtfx is the default referentiel, else take the first in array
+	- var referentiel = data.bdtfx ? 'bdtfx' : data.referentiels[0];
+	- var title = data._highlightResult[referentiel].scientific_name.value
+	- var href = data[referentiel].permalink
+	- var text = data._highlightResult[referentiel].common_name.value
 	h3.search-hit-title
-		a(href=data.bdtfx.permalink)!= title
+		a(href=href)!= title
+		each ref in data.referentiels
+			a.search-hit-tag(href=data[ref].permalink)= ref
 	if text
-		.search-hit-text!= text
+		.search-hit-text
+			!= text
 
 mixin search-hit-vegetation(data)
 	- var title = data._highlightResult.syntaxon.value
+	- var href = data.permalink
 	- var text = data._highlightResult.commonName.value
 	h3.search-hit-title
-		a(href=data.permalink)!= title
+		a(href=href)!= title
 	if text
 		.search-hit-text!= text
