+++ modules/search-hit/style.scss
@@ -1,18 +1,17 @@
 .search-hit {
-	& + & {
-		margin-top: 3.5rem;
-	}
+	margin-bottom: 3rem;
 
 	em {
 		background-color: $jaune-clair;
 		color: $noir;
+		display: inline-block;
 		font-style: normal;
 	}
 
 	&-title {
 		font-family: $ubuntu;
-		font-size: 1.9rem;
-		font-weight: $medium;
+		font-size: 1.8rem;
+		font-weight: $bold;
 		margin: 0;
 
 		&,
@@ -22,9 +21,26 @@
 
 		a {
 			text-decoration: none;
-			transition: color 0.2s ease;
+			transition: background-color 0.2s ease, color 0.2s ease;
 
 			&:hover {color: $turquoise-fonce;}
+		}
+
+		.search-hit-tag {
+			background-color: $vert-bouton;
+			border-radius: 0.2rem;
+			display: inline-block;
+			font-size: 1.25rem;
+			margin-left: 1rem;
+			padding: 0.25rem 0.5rem;
+			vertical-align: bottom;
+
+			&,
+			&:hover {
+				color: $blanc;
+			}
+
+			&:hover {background-color: $turquoise-fonce;}
 		}
 	}
