+++ package.json
@@ -57,6 +57,6 @@
     "tether": "^1.4.0",
     "url-loader": "^0.5.7",
     "velocity-animate": "^1.5.0",
-    "webpack": "^2.5.1"
+    "webpack": "^2.6.1"
   }
 }
